+++ demisto_sdk/commands/common/errors.py
@@ -16,14 +16,15 @@ ALLOWED_IGNORE_ERRORS = [
     'DS107',
     'GF102',
     'IF100', 'IF106', 'IF115', 'IF116',
-    'IN109', 'IN110', 'IN122', 'IN124', 'IN126', 'IN128', 'IN135', 'IN136', 'IN139', 'IN144', 'IN145',
+    'IN109', 'IN110', 'IN122', 'IN124', 'IN126', 'IN128', 'IN135', 'IN136', 'IN139', 'IN144', 'IN145', 'IN153', 'IN154',
     'MP106',
     'PA113', 'PA116', 'PA124', 'PA125', 'PA127', 'PA129',
     'PB104', 'PB105', 'PB106', 'PB110', 'PB111', 'PB112', 'PB114', 'PB115', 'PB116', 'PB107',
-    'RM100', 'RM102', 'RM104', 'RM106', 'RM108', 'RM 110',
+    'RM100', 'RM102', 'RM104', 'RM106', 'RM108', 'RM 110', 'RM112',
     'RP102', 'RP104',
     'SC100', 'SC101', 'SC105', 'SC106',
-    'IM111', 'RN112'
+    'IM111',
+    'RN112',
 ]
 
 # predefined errors to be ignored in partner/community supported packs even if they do not appear in .pack-ignore
@@ -56,6 +57,8 @@ ERROR_CODE = {
     "changed_pack_name": {'code': "BA114", 'ui_applicable': False, 'related_field': 'name'},
     "file_cannot_be_deleted": {'code': "BA115", 'ui_applicable': False, 'related_field': ''},
     "cli_name_and_id_do_not_match": {'code': "BA116", 'ui_applicable': False, 'related_field': 'cliName'},
+    "incorrect_from_to_version_format": {'code': "BA117", 'ui_applicable': False, 'related_field': ''},
+    "mismatching_from_to_versions": {'code': "BA118", 'ui_applicable': False, 'related_field': ''},
 
     # BC - Backward Compatible
     "breaking_backwards_subtype": {'code': "BC100", 'ui_applicable': False, 'related_field': 'subtype'},
@@ -105,6 +108,7 @@ ERROR_CODE = {
     "docker_not_on_the_latest_tag": {'code': "DO106", 'ui_applicable': True, 'related_field': 'dockerimage'},
     "non_existing_docker": {'code': "DO107", 'ui_applicable': True, 'related_field': 'dockerimage'},
     "dockerimage_not_in_yml_file": {'code': "DO108", 'ui_applicable': True, 'related_field': 'dockerimage'},
+    "deprecated_docker_error": {'code': "DO109", 'ui_applicable': True, 'related_field': 'dockerimage'},
 
     # DS - Descriptions
     "description_missing_in_beta_integration": {'code': "DS100", 'ui_applicable': False, 'related_field': ''},
@@ -231,7 +235,9 @@ ERROR_CODE = {
     'invalid_siem_integration_name': {'code': 'IN150', 'ui_applicable': True, 'related_field': 'display'},
     "empty_command_arguments": {'code': 'IN151', 'ui_applicable': False, 'related_field': 'arguments'},
     'invalid_defaultvalue_for_checkbox_field': {'code': 'IN152', 'ui_applicable': True, 'related_field': 'defaultvalue'},
-    'integration_is_deprecated_and_used': {'code': 'IN153', 'ui_applicable': True, 'related_field': 'deprecated'},
+    'not_supported_integration_parameter_url_defaultvalue': {'code': 'IN153', 'ui_applicable': False, 'related_field': 'defaultvalue'},
+    'missing_reliability_parameter': {'code': 'IN154', 'ui_applicable': False, 'related_field': 'configuration'},
+    'integration_is_deprecated_and_used': {'code': 'IN155', 'ui_applicable': True, 'related_field': 'deprecated'},
 
     # IT - Incident Types
     "incident_type_integer_field": {'code': "IT100", 'ui_applicable': True, 'related_field': ''},
@@ -298,6 +304,7 @@ ERROR_CODE = {
     "pack_metadata_missing_categories": {'code': "PA129", 'ui_applicable': False, 'related_field': ''},
     "wrong_version_format": {'code': "PA130", 'ui_applicable': False, 'related_field': ''},
     "pack_metadata_version_diff_from_rn": {'code': "PA131", 'ui_applicable': False, 'related_field': ''},
+    "pack_should_be_deprecated": {'code': "PA132", 'ui_applicable': False, 'related_field': ''},
 
     # PB - Playbooks
     "playbook_cant_have_rolename": {'code': "PB100", 'ui_applicable': True, 'related_field': 'rolename'},
@@ -341,6 +348,7 @@ ERROR_CODE = {
     "missing_readme_file": {'code': "RM109", 'ui_applicable': False, 'related_field': ''},
     "missing_commands_from_readme": {'code': "RM110", 'ui_applicable': False, 'related_field': ''},
     "error_uninstall_node": {'code': "RM111", 'ui_applicable': False, 'related_field': ''},
+    "invalid_readme_relative_url_error": {'code': "RM112", 'ui_applicable': False, 'related_field': ''},
 
     # RN - Release Notes
     "missing_release_notes": {'code': "RN100", 'ui_applicable': False, 'related_field': ''},
@@ -855,9 +863,15 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def not_supported_integration_parameter_url_defaultvalue(param, invalid_defaultvalue):
+        return f"The integration parameter {param} has defaultvalue set to {invalid_defaultvalue}. If possible, replace the http prefix with https."
+
+    @staticmethod
+    @error_code_decorator
     def is_valid_integration_file_path_in_folder(integration_file):
         return f"The integration file name: {integration_file} is invalid, " \
-               f"The integration file name should be the same as the name of the folder that contains it."
+               f"The integration file name and all the other files in the folder, should be the same as " \
+               f"the name of the folder that contains it."
 
     @staticmethod
     @error_code_decorator
@@ -882,9 +896,15 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def invalid_defaultvalue_for_checkbox_field(name: str):
-        return f"The defaultvalue checkbox of {name}'s filed is incorrect, " \
-               f"should be 'true' or 'false', a string which contains only lowercase letters.\n " \
-               f"e.g: defaultvalue: 'true'"
+        return f"The defaultvalue checkbox of the {name} field is invalid. " \
+               f"Use a boolean represented as a lowercase string, e.g defaultvalue: 'true'"
+
+    @staticmethod
+    @error_code_decorator
+    def missing_reliability_parameter(command: str):
+        return f'Missing "Reliability" parameter in the {command} reputation command.' \
+               f'Please add it to the YAML file.' \
+               f'For more information, refer to https://xsoar.pan.dev/docs/integrations/dbot#reliability-level'
 
     @staticmethod
     @error_code_decorator
@@ -965,8 +985,8 @@ class Errors:
     @classmethod
     @error_code_decorator
     def breaking_backwards_command_arg_changed(cls, commands_ls):
-        error_msg = "{}, Your updates to this file contains changes to a name or an argument of an existing command(s).\n" \
-            "Please undo you changes to the following command(s):\n".format(cls.BACKWARDS)
+        error_msg = "{}, Your updates to this file contains changes to a name or an argument of an existing " \
+                    "command(s).\nPlease undo you changes to the following command(s):\n".format(cls.BACKWARDS)
         error_msg += '\n'.join(commands_ls)
         return error_msg
 
@@ -983,6 +1003,11 @@ class Errors:
                f'Please create or update to an updated versioned image\n' \
                f'You can check for the most updated version of {docker_image_tag} ' \
                f'here: https://hub.docker.com/r/{docker_image_name}/tags'
+
+    @staticmethod
+    @error_code_decorator
+    def deprecated_docker_error(docker_image_name, deprecated_reason):
+        return f'The docker image {docker_image_name} is deprecated - {deprecated_reason}'
 
     @staticmethod
     @error_code_decorator
@@ -1120,8 +1145,9 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def invalid_image_name():
-        return "The image's file name is invalid - " \
-               "make sure the name looks like the following: <integration_name>_image.png"
+        return "The image's file name is invalid - make sure the name looks like the " \
+               "following: <integration_name>_image.png and that the integration_name is the same as the folder " \
+               "containing it."
 
     @staticmethod
     @error_code_decorator
@@ -1342,10 +1368,12 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
-    def content_entity_version_not_match_playbook_version(main_playbook, entities_names, main_playbook_version):
-        return f"Playbook {main_playbook} with version {main_playbook_version} uses {entities_names} " \
-               f"with a version that does not match the main playbook version. The from version of" \
-               f" {entities_names} should be {main_playbook_version} or lower."
+    def content_entity_version_not_match_playbook_version(
+        main_playbook: str, entities_names_and_version: str, main_playbook_version: str, content_sub_type: str
+    ):
+        return f"Playbook {main_playbook} with 'fromversion' {main_playbook_version} uses the following" \
+               f" {content_sub_type} with an invalid 'fromversion': [{entities_names_and_version}]. " \
+               f"The 'fromversion' of the {content_sub_type} should be {main_playbook_version} or lower."
 
     @staticmethod
     @error_code_decorator
@@ -1393,7 +1421,8 @@ class Errors:
     @error_code_decorator
     def invalid_description_name():
         return "The description's file name is invalid - " \
-               "make sure the name looks like the following: <integration_name>_description.md"
+               "make sure the name looks like the following: <integration_name>_description.md " \
+               "and that the integration_name is the same as the folder containing it."
 
     @staticmethod
     @error_code_decorator
@@ -1782,6 +1811,12 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def invalid_readme_relative_url_error(path):
+        return f'Relative urls are not supported within README. If this is not a relative url, please add ' \
+               f'an https:// prefix:\n{path}. '
+
+    @staticmethod
+    @error_code_decorator
     def invalid_readme_image_error(path: str, error_type: str, response: Optional[Response] = None):
         error = 'Error in readme image: '
         if response is not None:
@@ -2105,6 +2140,18 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def incorrect_from_to_version_format(incorrect_key: str):
+        return f"The format of the {incorrect_key} is incorrect\n" \
+               f"Please fix this so that it is in xx.xx.xx format and each member is a number only."
+
+    @staticmethod
+    @error_code_decorator
+    def mismatching_from_to_versions():
+        return 'The `fromversion` and `toversion` are not synchronizied\n' \
+               'It is must be fromversion <= toversion.'
+
+    @staticmethod
+    @error_code_decorator
     def integration_is_skipped(integration_id, skip_comment: Optional[str] = None):
         message = f"The integration {integration_id} is currently in skipped. Please add working tests and unskip."
         if skip_comment:
@@ -2282,7 +2329,7 @@ class Errors:
     @error_code_decorator
     def pack_metadata_version_diff_from_rn(pack_path, rn_version, pack_metadata_version):
         return f'There is a difference between the version in the pack metadata' \
-               f'file and the version of the latest release note.\nexpected latest release note to be {pack_metadata_version} '\
+               f'file and the version of the latest release note.\nexpected latest release note to be {pack_metadata_version} ' \
                f'instead found {rn_version}.\nTo fix the problem, try running `demisto-sdk update-release-notes -i {pack_path}`'
 
     @staticmethod
@@ -2354,3 +2401,12 @@ class Errors:
     @error_code_decorator
     def wizard_integrations_without_playbooks(integrations: set):
         return f'The following integrations are missing a set_playbook: {integrations}'
+
+    @staticmethod
+    @error_code_decorator
+    def pack_should_be_deprecated(pack_name: str):
+        return f'Pack {pack_name} should be deprecated, as all its integrations, playbooks and scripts are' \
+               f' deprecated.\nThe name of the pack in the pack_metadata.json should end with (Deprecated)\n' \
+               f'The description of the pack in the pack_metadata.json should be one of the following formats:\n' \
+               f'1. "Deprecated. Use <PACK_NAME> instead."\n' \
+               f'2. "Deprecated. <REASON> No available replacement."'