+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -114,8 +114,10 @@ class IntegrationValidator(ContentEntityValidator):
             self.is_valid_max_fetch_and_first_fetch(),
             self.is_valid_as_deprecated(),
             self.is_valid_parameters_display_name(),
+            self.is_valid_parameter_url_default_value(),
             self.is_mapping_fields_command_exist(),
             self.is_valid_integration_file_path(),
+            self.is_valid_py_file_names(),
             self.has_no_duplicate_params(),
             self.has_no_duplicate_args(),
             self.is_there_separators_in_names(),
@@ -149,6 +151,7 @@ class IntegrationValidator(ContentEntityValidator):
             self.is_valid_description(beta_integration=False),
             self.is_context_correct_in_readme(),
             self.verify_yml_commands_match_readme(is_modified),
+            self.verify_reputation_commands_has_reliability(is_modified),
             self.is_integration_deprecated_and_used()
         ]
 
@@ -1216,6 +1219,27 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN153')
+    def is_valid_parameter_url_default_value(self) -> bool:
+        """Verifies integration parameters default value is valid.
+
+        Returns:
+            bool: True if the default value of the url parameter uses the https protocol,
+            False otherwise.
+        """
+        configuration = self.current_file.get('configuration', {})
+        parameters_default_values = [(param.get('display'), param.get('defaultvalue')) for param in configuration if param.get('defaultvalue')]
+
+        is_valid = True
+        for param, defaultvalue in parameters_default_values:
+            if defaultvalue and isinstance(defaultvalue, str):
+                if defaultvalue.startswith('http:'):
+                    error_message, error_code = Errors.not_supported_integration_parameter_url_defaultvalue(param, defaultvalue)
+                    if self.handle_error(error_message, error_code, file_path=self.file_path):
+                        is_valid = False
+
+        return is_valid
+
     @error_codes('IN138,IN137')
     def is_valid_integration_file_path(self) -> bool:
         absolute_file_path = self.file_path
@@ -1240,6 +1264,33 @@ class IntegrationValidator(ContentEntityValidator):
                 error_message, error_code = Errors.is_valid_integration_file_path_in_folder(integration_file)
                 if self.handle_error(error_message, error_code, file_path=self.file_path):
                     return False
+
+        return True
+
+    @error_codes('IN137')
+    def is_valid_py_file_names(self):
+        # Gets the all integration .py files from the integration folder.
+        excluded_files = ['demistomock.py', 'conftest.py', 'CommonServerPython.py', 'CommonServerUserPython.py',
+                          '.vulture_whitelist.py']
+        files_to_check = get_files_in_dir(os.path.dirname(self.file_path), ['py'], False)
+        invalid_files = []
+        integrations_folder = os.path.basename(os.path.dirname(self.file_path))
+
+        for file_path in files_to_check:
+            file_name = os.path.basename(file_path)
+            if file_name not in excluded_files:
+                # The unittest has _test.py suffix whereas the integration only has the .py suffix
+                splitter = '_' if file_name.endswith('_test.py') else '.'
+                base_name = file_name.rsplit(splitter, 1)[0]
+
+                if integrations_folder != base_name:
+                    invalid_files.append(file_name)
+
+        if invalid_files:
+            error_message, error_code = Errors.is_valid_integration_file_path_in_folder(invalid_files)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                self.is_valid = False
+                return False
 
         return True
 
@@ -1597,7 +1648,7 @@ class IntegrationValidator(ContentEntityValidator):
         """
         Checks if there are commands that doesn't appear in the readme but appear in the .yml file
         Args:
-            is_modified (bool): Wether the given files are modified or not.
+            is_modified (bool): Whether the given files are modified or not.
 
         Return:
             bool: True if all commands are documented in the README, False if there's no README file in the expected
@@ -1627,7 +1678,31 @@ class IntegrationValidator(ContentEntityValidator):
 
         return is_valid
 
-    @error_codes('IN153')
+    @error_codes('IN154')
+    def verify_reputation_commands_has_reliability(self, is_modified: bool = False):
+        """
+        In case the integration has reputation command, ensure there is a reliability parameter.
+        Args:
+            is_modified (bool): Whether the given files are modified or not.
+
+        Return:
+            bool: True if there are no reputation commands or there is a reliability parameter
+             and False if there is at least one reputation command without a reliability parameter in the configuration.
+        """
+        if not is_modified:
+            return True
+        commands_names = [command.get('name') for command in self.current_file.get('script', {}).get('commands', [])]
+        yml_config_names = ' '.join([config.get('name') for config in self.current_file.get('configuration', {})])
+        for command in commands_names:
+            if command in REPUTATION_COMMAND_NAMES:
+                if 'reliability' in yml_config_names.lower():
+                    return True
+                error_message, error_code = Errors.missing_reliability_parameter(command)
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    return False
+        return True
+
+    @error_codes('IN155')
     def is_integration_deprecated_and_used(self):
         """
         Checks if there are commands that are deprecated and is used in other none-deprcated scripts / playbooks.