+++ demisto_sdk/commands/lint/commands_builder.py
@@ -19,32 +19,10 @@ from demisto_sdk.commands.lint.resources.pylint_plugins.xsoar_level_checker impo
 # Third party packages
 # Local imports
 
-excluded_files = ["CommonServerPython.py", "demistomock.py", "CommonServerUserPython.py", "conftest.py", "venv"]
+excluded_files = ["CommonServerPython.py", "demistomock.py", "CommonServerUserPython.py", "conftest.py", ".venv"]
 
 
-def get_python_exec(py_num: str, is_py2: bool = False) -> str:
-    """ Get python executable
-
-    Args:
-        py_num(str): Python version X.Y
-        is_py2(bool): for python 2 version, Set True if the returned result should have python2 or False for python.
-
-    Returns:
-        str: python executable
-    """
-    py_ver = parse(py_num).major
-    if py_ver < 3:
-        if is_py2:
-            py_str = "2"
-        else:
-            py_str = ""
-    else:
-        py_str = "3"
-
-    return f"python{py_str}"
-
-
-def build_flake8_command(files: List[Path], py_num: str) -> str:
+def build_flake8_command(files: List[Path]) -> str:
     """ Build command for executing flake8 lint check
         https://flake8.pycqa.org/en/latest/user/invocation.html
     Args:
@@ -55,10 +33,18 @@ def build_flake8_command(files: List[Path], py_num: str) -> str:
         str: flake8 command
     """
 
-    command = f"{get_python_exec(py_num)} -m flake8"
+    command = "flake8 "
+
+    # This is the same config used in `tox.ini` file in content
+    # We will probably want to use that in the future for this and all linters
+    command += '--ignore=W605,F403,F405,W503 '
+    command += '--exclude=_script_template_docker.py,./CommonServerPython.py,./demistomock.py '
+    command += '--max-line-length 130 '
+    command += '--per-file-ignores=nudge_external_prs.py:E231,E251,E999 '
+
     # Generating file patterns - path1,path2,path3,..
-    files_list = [str(file) for file in files]
-    command += ' ' + ' '.join(files_list)
+    files_list = [file.name for file in files]
+    command += ' '.join(files_list)
 
     return command
 
@@ -72,7 +58,7 @@ def build_bandit_command(files: List[Path]) -> str:
     Returns:
         str: bandit command
     """
-    command = "python3 -m bandit"
+    command = "bandit"
     # Reporting only issues with high and medium severity level
     command += " -ll"
     # Reporting only issues of a high confidence level
@@ -96,7 +82,7 @@ def build_bandit_command(files: List[Path]) -> str:
     return command
 
 
-def build_xsoar_linter_command(files: List[Path], py_num: str, support_level: str = "base") -> str:
+def build_xsoar_linter_command(files: List[Path], support_level: str = "base") -> str:
     """ Build command to execute with xsoar linter module
     Args:
         py_num(str): The python version in use
@@ -136,7 +122,7 @@ def build_xsoar_linter_command(files: List[Path], py_num: str, support_level: st
             for msg in checker_msgs_list:
                 message_enable += f"{msg},"
 
-    command = f"{get_python_exec(py_num, True)} -m pylint"
+    command = "pylint"
     # Excluded files
     command += f" --ignore={','.join(excluded_files)}"
     # Disable all errors
@@ -165,7 +151,7 @@ def build_mypy_command(files: List[Path], version: str, content_repo: Path = Non
     Returns:
         str: mypy command
     """
-    command = "python3 -m mypy"
+    command = "mypy"
     # Define python versions
     command += f" --python-version {version}"
     # This flag enable type checks the body of every function, regardless of whether it has type annotations.
@@ -184,6 +170,8 @@ def build_mypy_command(files: List[Path], version: str, content_repo: Path = Non
     command += " --allow-redefinition"
     # Get the full path to the file.
     command += " --show-absolute-path"
+    # Ignore site packages because running on host
+    command += " --no-site-packages"
     # Point cache to be .mypy_cache in the content repo
     command += f" --cache-dir={content_repo/'.mypy_cache' if content_repo else '/dev/null'}"
     # Generating path patterns - file1 file2 file3,..
@@ -193,7 +181,7 @@ def build_mypy_command(files: List[Path], version: str, content_repo: Path = Non
     return command
 
 
-def build_vulture_command(files: List[Path], pack_path: Path, py_num: str) -> str:
+def build_vulture_command(files: List[Path], pack_path: Path) -> str:
     """ Build command to execute with pylint module
         https://github.com/jendrikseipp/vulture
     Args:
@@ -204,7 +192,7 @@ def build_vulture_command(files: List[Path], pack_path: Path, py_num: str) -> st
     Returns:
        str: vulture command
     """
-    command = f"{get_python_exec(py_num)} -m vulture"
+    command = "vulture"
     # Excluded files
     command += f" --min-confidence {os.environ.get('VULTURE_MIN_CONFIDENCE_LEVEL', '100')}"
     # File to be excluded when performing lints check
@@ -212,8 +200,8 @@ def build_vulture_command(files: List[Path], pack_path: Path, py_num: str) -> st
     # Whitelist vulture
     whitelist = Path(pack_path) / '.vulture_whitelist.py'
     if whitelist.exists():
-        command += f" {whitelist}"
-    files_list = [str(item) for item in files]
+        command += f" {whitelist.name}"
+    files_list = [file.name for file in files]
     command += " " + " ".join(files_list)
     return command
 
@@ -227,7 +215,7 @@ def build_pylint_command(files: List[Path], docker_version: Optional[str] = None
     Returns:
        str: pylint command
     """
-    command = "python -m pylint"
+    command = "pylint"
     # Excluded files
     command += f" --ignore={','.join(excluded_files)}"
     # Prints only errors
@@ -267,7 +255,7 @@ def build_pytest_command(test_xml: str = "", json: bool = False, cov: str = "") 
     Returns:
         str: pytest command
     """
-    command = "python -m pytest -ra"
+    command = "pytest -ra"
     # Generating junit-xml report - used in circle ci
     if test_xml:
         command += " --junitxml=/devwork/report_pytest.xml"